fix warnings

+++ main.cpp
@@ -331,10 +331,10 @@ int main(int argc, const char * argv[]) {
   }
 
   auto durationPoints = boost::chrono::high_resolution_clock::now() - startPoints;
-  printf("All points read in %d seconds || %02d:%02d:%02d\n", 
-    (int)boost::chrono::duration_cast<boost::chrono::seconds>(durationPoints).count(),
-    (int)boost::chrono::duration_cast<boost::chrono::hours>(durationPoints).count(),
-    (int)boost::chrono::duration_cast<boost::chrono::minutes>(durationPoints).count() % 60,
+  printf("All points read in %ld seconds || %02d:%02d:%02d\n",
+    boost::chrono::duration_cast<boost::chrono::seconds>(durationPoints).count(),
+    boost::chrono::duration_cast<boost::chrono::hours>(durationPoints).count(),
+    boost::chrono::duration_cast<boost::chrono::minutes>(durationPoints).count() % 60,
     (int)boost::chrono::duration_cast<boost::chrono::seconds>(durationPoints).count() % 60
   );
 
@@ -412,14 +412,14 @@ int main(int argc, const char * argv[]) {
   }
   of.close();
 
-  printf("File written in %d ms\n", std::clock() - startFileWriting);
+  printf("File written in %ld ms\n", (int)(std::clock() - startFileWriting));
 
   //-- bye-bye
   auto duration = boost::chrono::high_resolution_clock::now() - startTime;
-  printf("Successfully terminated in %d seconds || %02d:%02d:%02d\n", 
-    (int)boost::chrono::duration_cast<boost::chrono::seconds>(durationPoints).count(),
-    (int)boost::chrono::duration_cast<boost::chrono::hours>(durationPoints).count(),
-    (int)boost::chrono::duration_cast<boost::chrono::minutes>(durationPoints).count() % 60,
+  printf("Successfully terminated in %ld seconds || %02d:%02d:%02d\n",
+    boost::chrono::duration_cast<boost::chrono::seconds>(durationPoints).count(),
+    boost::chrono::duration_cast<boost::chrono::hours>(durationPoints).count(),
+    boost::chrono::duration_cast<boost::chrono::minutes>(durationPoints).count() % 60,
     (int)boost::chrono::duration_cast<boost::chrono::seconds>(durationPoints).count() % 60
   );
   return 1;
